Use common dashboard component
