+++ src/views/Project/index.jsx
@@ -5,20 +5,17 @@ import dotProp from 'dot-prop-immutable';
 import { mergeAll, memoizeWith } from 'ramda';
 import uniqBy from 'lodash.uniqby';
 import Typography from '@material-ui/core/Typography';
-import { NavLink } from 'react-router-dom';
 import { withStyles } from '@material-ui/core/styles';
-import IconButton from '@material-ui/core/IconButton';
 import ExpansionPanel from '@material-ui/core/ExpansionPanel';
 import ExpansionPanelDetails from '@material-ui/core/ExpansionPanelDetails';
 import ExpansionPanelSummary from '@material-ui/core/ExpansionPanelSummary';
 import Markdown from 'react-markdown';
-import ArrowLeftIcon from 'mdi-react/ArrowLeftIcon';
 import ChevronDownIcon from 'mdi-react/ChevronDownIcon';
 import projects from '../../data/loader';
 import Spinner from '../../components/Spinner';
-import AppBar from '../../components/AppBar';
 import ErrorPanel from '../../components/ErrorPanel';
 import TasksTable from '../../components/TasksTable';
+import Dashboard from '../../components/Dashboard';
 import issuesQuery from './issues.graphql';
 import bugsQuery from '../bugs.graphql';
 import commentsQuery from '../comments.graphql';
@@ -123,28 +120,6 @@ const tagReposMapping = repositories =>
 )
 @withApollo
 @withStyles(theme => ({
-  root: {
-    background: theme.palette.background.default,
-  },
-  header: {
-    height: 60,
-    paddingBottom: theme.spacing.unit,
-  },
-  container: {
-    marginTop: 60,
-    padding: 2 * theme.spacing.unit,
-    minHeight: `calc(100vh - 180px - ${3 * theme.spacing.unit}px)`,
-  },
-  link: {
-    textDecoration: 'none',
-    position: 'absolute',
-    '& svg': {
-      fill: theme.palette.common.white,
-    },
-  },
-  title: {
-    padding: '0 41px',
-  },
   spinner: {
     marginTop: 3 * theme.spacing.unit,
   },
@@ -368,56 +343,38 @@ export default class Project extends Component {
       [];
 
     return (
-      <div className={classes.root}>
-        <AppBar position="absolute" className={classes.header}>
-          <IconButton
-            aria-label="Back"
-            className={classes.link}
-            component={NavLink}
-            to="/">
-            <ArrowLeftIcon />
-          </IconButton>
-          <Typography
-            className={classes.title}
-            variant="display1"
-            align="center"
-            noWrap>
-            {project.name}
-          </Typography>
-        </AppBar>
-        <div className={classes.container}>
-          {project.introduction && (
-            <ExpansionPanel>
-              <ExpansionPanelSummary expandIcon={<ChevronDownIcon />}>
-                <Typography variant="headline">Project Introduction</Typography>
-              </ExpansionPanelSummary>
-              <ExpansionPanelDetails>
-                <Typography variant="body1">
-                  <Markdown
-                    source={project.introduction}
-                    renderers={{ link: this.linkRenderer }}
-                  />
-                </Typography>
-              </ExpansionPanelDetails>
-            </ExpansionPanel>
-          )}
-          {githubData &&
-            githubData.error && <ErrorPanel error={githubData.error} />}
-          {bugzillaData &&
-            bugzillaData.error && <ErrorPanel error={bugzillaData.error} />}
-          {error && <ErrorPanel error={error} />}
-          {loading && <Spinner className={classes.spinner} />}
-          {!loading && (
-            <TasksTable
-              onBugInfoClick={this.handleBugInfoClick}
-              items={uniqBy(
-                [...issues, ...goodFirstBugs, ...mentoredBugs],
-                'summary'
-              )}
-            />
-          )}
-        </div>
-      </div>
+      <Dashboard withSidebar={false}>
+        {project.introduction && (
+          <ExpansionPanel>
+            <ExpansionPanelSummary expandIcon={<ChevronDownIcon />}>
+              <Typography variant="headline">Project Introduction</Typography>
+            </ExpansionPanelSummary>
+            <ExpansionPanelDetails>
+              <Typography variant="body1">
+                <Markdown
+                  source={project.introduction}
+                  renderers={{ link: this.linkRenderer }}
+                />
+              </Typography>
+            </ExpansionPanelDetails>
+          </ExpansionPanel>
+        )}
+        {githubData &&
+          githubData.error && <ErrorPanel error={githubData.error} />}
+        {bugzillaData &&
+          bugzillaData.error && <ErrorPanel error={bugzillaData.error} />}
+        {error && <ErrorPanel error={error} />}
+        {loading && <Spinner className={classes.spinner} />}
+        {!loading && (
+          <TasksTable
+            onBugInfoClick={this.handleBugInfoClick}
+            items={uniqBy(
+              [...issues, ...goodFirstBugs, ...mentoredBugs],
+              'summary'
+            )}
+          />
+        )}
+      </Dashboard>
     );
   }
 }
